refactor: 게시판페이지 parallel route 적용, 랜딩페이지 그룹화
