+++ src/app/(common)/(board)/boards/(lists)/_components/BestItem.module.scss
