+++ src/app/(common)/(board)/boards/(lists)/_components/BoardItem.module.scss
