+++ src/app/(common)/(board)/boards/(lists)/page.tsx
@@ -1,0 +1,3 @@
+export default function BoardsPage() {
+  return <></>;
+}
