+++ src/app/(common)/(landing)/_components/Banner.module.scss
