+++ src/app/(common)/(landing)/_components/Banner.tsx
