+++ src/app/(common)/(landing)/_components/landingContents.ts
