+++ src/app/(common)/(market)/items/(lists)/_components/ProductFilter.tsx
@@ -1,12 +1,12 @@
 "use client";
 
+import useRecentSearch from "@/hooks/useRecentSearch";
 import useParams from "@/hooks/useParams";
 import { Recent, Search } from "@/components/Search";
-import { Select } from "@/components/ui";
-import styles from "./BoardFilter.module.scss";
-import useRecentSearch from "@/app/(common)/(market)/items/_components/useRecentSearch";
+import { Button, Select } from "@/components/ui";
+import styles from "./ProductFilter.module.scss";
 
-export default function BoardFilter() {
+export default function ProductFilter() {
   const { searchParams, handleParams } = useParams();
 
   const keyword = searchParams.get("keyword") || "";
@@ -32,8 +32,8 @@ export default function BoardFilter() {
   }
 
   return (
-    <div className={styles.filter}>
-      <div className={styles.keyword}>
+    <>
+      <div className={styles.search}>
         <Recent
           title="최근검색"
           data={recentSearch}
@@ -46,20 +46,25 @@ export default function BoardFilter() {
             onChange={handleSearchChange}
             onSubmit={handleSearchSubmit}
             onClear={handleSearchClear}
-            placeholder="검색어를 입력해주세요."
+            placeholder="검색할 상품을 입력해주세요"
           />
         </Recent>
       </div>
-      <div className={styles.order}>
+      <div className={styles.button}>
+        <Button href="/addItem" size="sm">
+          상품 등록하기
+        </Button>
+      </div>
+      <div className={styles.select}>
         <Select
           value={orderBy}
           onChange={handleOrderBy}
           options={[
             { value: "recent", label: "최신순" },
-            { value: "like", label: "좋아요순" },
+            { value: "favorite", label: "좋아요순" },
           ]}
         />
       </div>
-    </div>
+    </>
   );
 }
