+++ src/app/(common)/(market)/items/(lists)/layout.tsx
@@ -1,0 +1,20 @@
+import { PageWrapper } from "@/components/Page";
+import { ReactNode } from "react";
+
+export default function ListLayout({
+  all,
+  best,
+  children,
+}: {
+  all: ReactNode;
+  best: ReactNode;
+  children: ReactNode;
+}) {
+  return (
+    <PageWrapper>
+      {best}
+      {all}
+      {children}
+    </PageWrapper>
+  );
+}
