+++ src/app/(common)/(market)/items/(lists)/page.tsx
@@ -1,0 +1,3 @@
+export default function ItemsPage() {
+  return <></>;
+}
